Update EditTeam to use V1 endpoint
We always send *all* the data. Even though it's a PATCH endpoint, our existing EditTeam interface makes it easier to send everything. We could re-evaluate this later, but it's probably not going to making a meaningful difference.

I also had to change the field name to match the API so that validation errors are automatically displayed properly