VEGA-641: Use v1 endpoint to update teams (#82)
* Remove references to prototype

* Update EditTeam to use V1 endpoint

We always send *all* the data. Even though it's a PATCH endpoint, our existing EditTeam interface makes it easier to send everything. We could re-evaluate this later, but it's probably not going to making a meaningful difference.

I also had to change the field name to match the API so that validation errors are automatically displayed properly

* Remove reference to now-deleted config file

* Switch method to PUT

The upstream API will support PUT requests, not PATCH

* Update state for edit team tests

We need a team to edit, so reuse the existing "Supervision team with members exists" state

* Update label IDs

* Give test teams unique names

Otherwise the provider tests fail because each test uses a separate team entity

* Fix edit team test data

Only use member ID 23, because it will be handled bby pact proxy

Replace `teamTypeAlreadyInUse` key with `error`. The key doesn't matter to us, it's a relic, but Pact insists that we match the API exactly.

* Create slice with correct length

We can define the length up-front rather than leaving it variable and extending it with append-ings