Use Veracode uploadandscan to scan pushes to build
This is a longer process than pipeline scanning, and only one scan can run at any one time, so it should only be run on merges to main.

It creates a zip file [based on Veracode's requirements](https://help.veracode.com/r/compilation_go), uploads it, scans it and stores the results in Veracode.