Fix docker tagging
`github.head_ref` is only set on pull request events, so we should handle push events differently. There isn't an alternative variable, but we can only push to the main branch so we know that will be the branch name.

Also push `--all-tags` to ensure both `latest` and the semver tag are pushed on main builds.