Add dependabot alternative for verifying pacts
If the PR was written by dependabot, we can't write to Pact Broker to verify the generated contract. Instead, just compare it to the latest version and check there are no new changes.

We allow differences where things are _missing_ from the generated file because the remote version has links and IDs.